+++ internal/app/exit_code.go
@@ -1,6 +1,6 @@
 package app
 
-import "multi-pocketbase-ui/internal/apperr"
+import "github.com/jiseop121/pbdash/internal/apperr"
 
 func MapErrorToExitCode(err error) int {
 	return apperr.ExitCode(err)